refactor(mobile): assets/images を用途別ディレクトリに整理
平置きされていた画像を backgrounds / icons / illustrations / logos に分類し、
typo (hilight) や `-` / `_` 混在の命名を snake_case に統一して見通しを良くした。
あわせて BreakScreen の prettier 整形漏れも解消。
